\name{searMETEO}
\alias{searMETEO}
\title{
Search for the variable table of a weather station
}
\description{
This function helps for search the variable table of 10-minutes data. 
}
\usage{
searMETEO(x)
}
\arguments{
  \item{x}{
Name of weather station. 
}
}

\value{
Data.Frame. 
}
\author{
Dominic Roye
}
\examples{
searMETEO("Santiago")
  }

